# docs/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The PaymentFlow payments API. This document describes the public `/v1` surface — the endpoints PaymentFlow versions, deprecates on a published timeline, and supports for integrators. Endpoints not described here are internal and carry no compatibility promise.

    Amounts are integers in the currency's minor unit (`amountMinor`): `1000` in `USD` is $10.00. There are no floating-point amounts anywhere in this API.

    Every mutating request accepts an `Idempotency-Key` header and every list is cursor-paginated; both are described on the operations that use them.
  license:
    name: Proprietary
  title: PaymentFlow API
  version: 2026-08-01
servers:
- url: https://api.paymentflow.dev
  description: PaymentFlow API
security:
- SecretKey: []
paths:
...
  /v1/events:
    get:
      description: |-
        Returns everything that has happened on your account, most recent first, cursor-paginated.

        This is the same feed your webhook endpoints are delivered from, which makes it the recovery path when they are not: an endpoint that was down, misconfigured, or auto-disabled loses deliveries, not events. Poll here to catch up, and reconcile by `id` — an event's `evt_` id is identical in both places.

        Ordered by when things **occurred**, not when they were recorded, so a redelivery cannot reorder your feed.
      operationId: listEvents
      parameters:
      - description: "How many objects to return, 1–100. Values above the ceiling are clamped to it rather than rejected, so a client asking for more simply receives the maximum. Defaults to 25."
        in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
      - description: A cursor for pagination. Pass the `nextCursor` from the previous response to fetch the following page. The value is opaque and signed — treat it as a token and never parse or construct one.
        in: query
        name: starting_after
        required: false
        schema:
          type: string
      - description: Return only events of this type. Must be a name from the event vocabulary — a typo is a `400` naming the valid values rather than an empty page.
        example: payment.captured
        in: query
        name: type
        required: false
        schema:
          type: string
      - description: "Return only objects created at or after this instant, as RFC 3339 (`2026-07-29T00:00:00Z`)."
        in: query
        name: created_after
        required: false
        schema:
          type: string
          format: date-time
      - description: "Return only objects created at or before this instant, as RFC 3339 (`2026-07-29T23:59:59Z`)."
        in: query
        name: created_before
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageEventResponse"
          description: "A page of events, most recent first."
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "The query could not be understood: a tampered or malformed `starting_after` cursor, a `limit` that is not positive, or a filter value outside the vocabulary this endpoint accepts. A rejected filter is deliberately an error rather than an empty page — an empty page is something you would then have to explain to yourself."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: List events
      tags:
      - Events
  /v1/events/{id}:
    get:
      description: "Returns one event by the `evt_` id you received in a webhook body. This is how a receiver confirms that a delivery it was sent is genuine and still says what it said — the id in the webhook and the id here are the same value, derived from the same source."
      operationId: getEvent
      parameters:
      - description: The `evt_` id of the event to retrieve.
        example: evt_9f2c1e7a4b8d4c3e8a1d2b4f6a8c05d1
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
          description: The event.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: The id is not a well-formed `evt_` identifier.
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "No such event. Also returned for an event that exists but belongs to another merchant, to the other mode, or is internal to the platform and not part of your feed — never `403`, which would confirm it exists."
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Retrieve an event
      tags:
      - Events
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: "A stable, machine-readable identifier for this specific failure, such as `PAYMENT_NOT_CAPTURABLE`. The set of codes grows by policy, which is why `type` and not this is what a client should switch on. Every code is catalogued in docs/ERRORS.md."
        correlationId:
          type: string
          description: "Identifies the whole distributed trace this call belongs to, which may span several services. Broader than `requestId`."
        docUrl:
          type: string
          description: Where to read about this specific code.
        errors:
          type: array
          description: "Field-level validation failures, when the request failed validation in more than one place."
          items:
            $ref: "#/components/schemas/ApiFieldError"
        message:
          type: string
          description: "A human-readable explanation. Safe to log; not intended to be shown to your customers, and never programmatically parsed."
        param:
          type: string
          description: "The single offending parameter, when there is exactly one. Multi-field validation failures use `errors` instead."
        path:
          type: string
          description: The request path that produced the error.
        requestId:
          type: string
          description: Identifies this one HTTP call. **Quote this in a support request** — it is on every log line and every row of your request log.
        status:
          type: integer
          format: int32
          description: "The HTTP status code, repeated in the body so a logged response is self-contained."
        timestamp:
          type: string
          format: date-time
          description: "When the error occurred, as RFC 3339."
        type:
          type: string
          description: "The error's classification, and **the field to branch on**. This is a small closed set an SDK maps to exception classes: `authentication_error`, `permission_error`, `invalid_request_error`, `idempotency_error`, `rate_limit_error`, `api_error`."
    ApiFieldError:
      type: object
      properties:
        field:
          type: string
          description: The request field that failed validation.
        message:
          type: string
          description: Why it failed. The rejected value is deliberately never echoed back.
...
    CursorPageEventResponse:
      type: object
      properties:
        data:
          type: array
          description: "The objects on this page, most recent first."
          items:
            $ref: "#/components/schemas/EventResponse"
        hasMore:
          type: boolean
          description: Whether more objects exist after this page. A cursor page reports no total count — that would cost a second full count query on every request — and this is the only question a paginating client needs.
        nextCursor:
          type: string
          description: "Pass this as `starting_after` to fetch the next page. Absent when `hasMore` is false. Opaque and signed: treat it as a token, never parse or construct one."
        object:
          type: string
          description: "Always `list`. A constant discriminator, so a client deserializing a response can branch on a field rather than on the endpoint it called."
...
    EventResponse:
      type: object
      properties:
        created:
          type: string
          format: date-time
          description: "When the event occurred — not when it was recorded. Events are ordered by this, so a redelivery cannot reorder your feed."
        data:
          type: object
          additionalProperties: true
          description: "The event payload — the object the event happened to, as it was at the time. Its shape depends on `type` and is the same body your webhook endpoint received. Stored verbatim: audit-service records what it was given rather than a typed view it would then have to keep in step."
        id:
          type: string
          description: "Unique identifier for this event, `evt_` followed by 32 hex characters. **Byte-identical to the `id` in the webhook body you received for the same event**, so a webhook can be reconciled against this log without storing anything extra."
          example: evt_9f2c1e7a4b8d4c3e8a1d2b4f6a8c05d1
        mode:
          type: string
          description: Whether this event describes `test` or `live` activity.
          enum:
          - test
          - live
          example: test
        object:
          type: string
          description: Always `event`. The discriminator that identifies this object out of context.
          example: event
        type:
          type: string
          description: "What happened, from the frozen event vocabulary — `payment.created`, `payment.authorized`, `payment.captured`, `payment.refunded`, `payment.voided`, `payment.failed`. **New types may be added without a new API revision**, so treat an unrecognised type as one you do not handle."
          example: payment.captured
...
  securitySchemes:
    SecretKey:
      bearerFormat: sk
      description: |-
        Your secret API key, sent as `Authorization: Bearer sk_test_...`.

        The key determines both *whose* data you see and *which mode* you see it in: an `sk_test_` key reads and writes test data only, an `sk_live_` key live data only. No header, parameter, or body field can change either — to work in the other mode, use the other key.
      scheme: bearer
      type: http